+++ CHANGELOG.md
@@ -3,14 +3,16 @@ All important changes to the decitala package will be documented here.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.5.0] - January 12, 2020 (Kent, CT)
+## [v0.5.0] - January 16, 2020 (Kent, CT)
 ### Added
 - Added a ``create_fragment_database`` function in ``database.py`` that holds name/ql data & ratio/difference equivalents in the decitala and greek metric databases (including equivalents _across_ databases). This data is available in ``fragment_database.db`` in the databases directory. 
 - Decitala and GreekFoot objects now have an ``equivalents`` method that return equivalents (based on the ``rep_type``) in the fragment corpus.
+- TODO: Created a Travis-CI account for continuous integration. 
 
 ### Changed
 - Decitala and Greek Metric instantiation now relies on SQL database instead of ``os.listdir``.
-- FragmentTree instantiation now relies on SQL database instead of os.listdir. 
+- TODO: FragmentTree default input is now a list. To create from a ``frag_type``, use the class method ``FragmentTree.from_frag_type``. 
+- FragmentTree instantiation now relies on SQL database instead of os.listdir.
 
 ## [v0.4.4] - January 10, 2020 (Kent, CT)
 ### Added